Analyze the replacement in the same context
In order to write (partial f x y ...) x, y and soforth need to be in
scope which we only know they can be at the point where f was
eliminated. Thus we have to pull the env out of f and reuse it to
analyze the replacement term. Nicola OK'd it and it works so it stays.