Implement the closure of closures for letfn
Correctly hoisting letfn requires that as the letfn terms are presumably
mutually recursive and may contain closures that the  closed-overs set
for any given letfn term when hoisted is not the naive closed-overs set
but rather is the fixed point merge of the closed-overs set of all
reachable expressions including the hoists of the other letfn
functions.

This patch makes the appropriate modifications to the handling of letfn
so that when letfn bodies are rewritten into partials the generated
rewritten functions and partials will correctly reflect the subset of
the hoisted environment which must be passed as arguments.