Fix #1
This patch fixes issue #1 by reworking check-vars to examine the
dependency set of vars in a def rather than checking the def'd var
itself. Note that this means check-vars should be run after tree-shaking
to escape errors in unreachable code (for instance the definition of
c.c/eval which is self-reaching).