Add core-redefs
This namespace provides custom definitions of the ns macros and some
other macros related to class generation in the Clojure core. A future
patch to load will have to provide these as bindings during code loading
in order for them to take effect.

Note that these bindings break the Clojure core, meaning that until
defmethod and soforth are supported by Oxcart it will not be able to
compile the core itself and will be restricted entirely to using the
existing runtime.