Prevent lifting of single fns
Lifted arities are tagged ^:single and ^:static. The former is true by
definition, the latter is true by virtue of the fact that a generated
fn cannot be taken as a value by the programer, is not taken as a
value by the compiler and thus is not taken as a value. The checking
of metadata for ^:single prevents the lifting of single arity emitted
fns. As normal fns are presumed to lack the ^:single metadata, normal
fns will be lifted into N ^:single fns and a "dispatch" rewritten core
fn. These cannot be re-lifted as the "dispatch" redefinition is part
of the munge table which is checked prior to lifting.

Thus all fns are either ^:single ^:static, or are munged and this
operation is now fixed point safe.

Fix #19